+++ .gitmodules
@@ -31,3 +31,6 @@
 [submodule "deps/azure-ctest"]
 	path = deps/azure-ctest
 	url = https://github.com/Azure/azure-ctest.git
+[submodule "deps/azure-iot-security-sdk"]
+	path = deps/azure-iot-security-sdk
+	url = https://github.com/Azure/azure-iot-security-sdk-c.git
