+++ iothub_client/inc/iotsecurity_client.h
